+++ CHANGELOG.md
@@ -8,10 +8,12 @@
 - Improved match caching performance
 - Added "meshes" blank folder to archive to make mod managers happy
 - Added a default blocklist rule for other LOD meshes
+- Added errors for failing texture hook patchers
 - Fixed system locale messing with certain characters in textures
 - Fixed texture names in plugins for turkish locales
 - Fixed backslashes in zip files for linux users
 - Fixed crash from certain kinds of log messages
+- Fixed vanilla SSS patcher not generating some texture maps
 
 ## [0.9.3] - 2025-10-18